Create myAccount mixin
